updated create user step
